Lot of modifications
1) Added interrupt.c and .h files. These files will handle interrupts from all peripherals from GB.

2) Modified cpu registers struct. Now it's a typedef (according to new project pattern) and the variable instanced is GB_cpu.

3) Everything written in gpu.c was commented and all lines involving gpu (except vram and oam arrays) in mmu were also commented.

4) Fixed error in one function prototype from timer.c/h.

+++ GB Emulator/mmu.h
@@ -55,6 +55,6 @@ void reset();
 uint8_t rdByte(uint16_t addr);
 uint16_t rdWord(uint16_t addr);
 void wrByte(uint16_t addr, uint8_t val);
-void wrWord(uint16_t addr, uint8_t val);
+void wrWord(uint16_t addr, uint16_t val);
 
 #endif
